requirements.txt
easydict==1.10
google_api_python_client==2.87.0
pandas==1.3.5
protobuf==4.23.2
